refactor(T1c PM): impl-split SFPipeAutoConnectManager across 2 .cpp [BUILD-PASSED]
Slice PM. Pure impl-split (one class, 2 .cpp): connection evaluation / manifold-connector /
pipe-child spawning / floor-hole pipes move to SFPipeAutoConnectManager_Spawn.cpp (1,266 lines).
SFPipeAutoConnectManager.cpp 2,789 -> 1,550. Static class-member PipeChildCounter definition stays
in the core TU (accessed as a class member from both). Zero behavior change. Shipping PackagePlugin
clean (ExitCode=0). Solo lane.
